+++ src/main/java/kr/kro/airbob/domain/discountPolicy/AccommodationDiscountPolicy.java
@@ -8,7 +8,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kr.kro.airbob.common.BaseEntity;
-import kr.kro.airbob.domain.accommodation.Accommodation;
+import kr.kro.airbob.domain.accommodation.entity.Accommodation;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
