+++ src/main/java/kr/kro/airbob/domain/image/AccommodationImage.java
@@ -4,8 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.kro.airbob.domain.accommodation.Accommodation;
-import kr.kro.airbob.domain.image.Image;
+import kr.kro.airbob.domain.accommodation.entity.Accommodation;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
