+++ src/main/java/kr/kro/airbob/domain/pricePolicy/PricePolicy.java
@@ -10,7 +10,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kr.kro.airbob.common.BaseEntity;
-import kr.kro.airbob.domain.accommodation.Accommodation;
+import kr.kro.airbob.domain.accommodation.entity.Accommodation;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
