+++ src/main/java/kr/kro/airbob/domain/review/Review.java
@@ -8,7 +8,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kr.kro.airbob.common.BaseEntity;
-import kr.kro.airbob.domain.accommodation.Accommodation;
+import kr.kro.airbob.domain.accommodation.entity.Accommodation;
 import kr.kro.airbob.domain.member.Member;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
